Added tests to check if feature flag is properly being evaluated before adding $status operation.
